fix: changed id from integer to long

+++ planning server/src/main/java/com/example/demo/models/Appeal.java
@@ -22,7 +22,7 @@ public class Appeal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "appeal_id")
-    private Integer appealId;
+    private Long appealId;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     @Column(name = "registration_date")
